Added a method too AppDelegate to populate the databases with fake user and transactions, so I can find bugs. Hooked up the user search to the database. fixed some bugs

+++ StormGolf/HDItemManager.h
@@ -10,7 +10,7 @@
 
 @interface HDItem : NSObject <NSCoding>
 @property (nonatomic, copy) NSString *itemDescription;
-@property (nonatomic, assign) NSUInteger itemCost;
+@property (nonatomic, assign) double itemCost;
 @end
 
 @interface HDItemManager : NSObject
